pyproject.toml
[project]
name = "dosaurus-hoster-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
